+++ docs/.vitepress/theme/config/teekConfig.ts
@@ -20,6 +20,7 @@ export const teekBlogConfig: TeekConfig = {
     imageViewer: {
       hideOnClickModal: true,
     },
+    coverImgMode: "full",
   },
   footerInfo: {
     customHtml: `<span id="runtime"></span>`, // 需要搭配 .vitepress/theme/helper/useRuntime.ts 使用
@@ -37,7 +38,7 @@ export const teekBlogConfig: TeekConfig = {
   },
   blogger: {
     avatar: "https://testingcf.jsdelivr.net/gh/Kele-Bingtang/static/user/avatar1.png",
-    shape: "square",
+    shape: "circle-rotate",
     name: "天客",
     slogan: "朝圣的使徒，正在走向编程的至高殿堂！",
   },